Web Fundamentals JS Varibable contern
Finished with varaible content addition.

Modified code element style to match style of when code is in a pre tag (black texst on white background. The defauult style was red font against the page background color. The change makes the inline code (not within a pre node) more readable.

Updated images for varaibles.