Move FontAwesome Script to head
Move scirpt for FontAwesome back to head element. This script is necessary for loading the GitHub and LinkedIn icons on the About | Find Me menu. I had moved the reference to the bottom due to page audit recommendations, but it seemed the menu icons were slower to load--noticeable lag in dispay when clicking the menu item. My hope is to have the icon display seem more seamless by starting the load process earlier. Asynch allows parallel processing.